// src/Blazor.Concurrency.Core/assets/ts/generated/http-request-data.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export interface HttpRequestData {
    url: string;
    method: string;
    body: Object | null;
    headers: Record<string, any> | null;
    maxRetries: number;
    timeoutMs: number | null;
}